+++ Runtime/UI/Components/FloatField.cs
@@ -30,16 +30,13 @@ namespace Unity.AppUI.UI
         /// <inheritdoc cref="NumericalField{T}.ParseValueToString"/>
         protected override string ParseValueToString(float val)
         {
-            // If formatString is percent, we need to divide by 100
-            if (UINumericFieldsUtils.IsPercentFormatString(formatString))
-            {
-                val /= 100f;
-                if (highValue == 1f)
-                    Debug.LogWarning("High value is set to 1, but format string is a percent format string.\n" +
-                        "Please set high value to 100 instead.");
-            }
-            
             return val.ToString(formatString, CultureInfo.InvariantCulture.NumberFormat);
+        }
+        
+        /// <inheritdoc cref="NumericalField{T}.ParseRawValueToString"/>
+        protected override string ParseRawValueToString(float val)
+        {
+            return val.ToString(CultureInfo.InvariantCulture.NumberFormat);
         }
 
         /// <inheritdoc cref="NumericalField{T}.AreEqual"/>