+++ Runtime/UI/Components/RangeSlider.cs
@@ -689,17 +689,8 @@ namespace Unity.AppUI.UI
         /// <inheritdoc cref="BaseSlider{Vector2,Single}.ParseValueToString"/>
         protected override string ParseValueToString(Vector2 val)
         {
-            // If formatString is percent, we need to divide by 100
-            if (UINumericFieldsUtils.IsPercentFormatString(formatString))
-            {
-                val /= 100f;
-                if (highValue == 1f)
-                    Debug.LogWarning("High value is set to 1, but format string is a percent format string.\n" +
-                        "Please set high value to 100 instead.");
-            }
-            
-            return $"[{val.x.ToString(formatString, CultureInfo.InvariantCulture)} - " +
-                $"{val.y.ToString(formatString, CultureInfo.InvariantCulture)}]";
+            return $"[{val.x.ToString(formatString, CultureInfo.InvariantCulture.NumberFormat)} - " +
+                $"{val.y.ToString(formatString, CultureInfo.InvariantCulture.NumberFormat)}]";
         }
 
         /// <inheritdoc cref="BaseSlider{Vector2,Single}.SliderLerpUnclamped"/>